requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.1
scikit-learn>=1.5.0
pydantic>=2.8.0
pytest>=8.3.0
